+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to a section with the version name.
 * Prevent manual submission of `Sales Invoice Additional Fields` where user can submit with the intent of resending it ([Issue](https://github.com/lavaloon-eg/ksa_compliance/issues/233))
   * It should only be submitted by the hourly job or the sync page.
 
+* Add revocation option for zatca business settings and CSID.
 ## 0.47.0
 
 * Support displaying `Return Against Additional References` in `ZATCA Phase 2 print format`.